refactor(client): modularize save note dialog into separate components
Create dedicated header component for save note dialog:
  - Extracted header content into a separate component for better reusability and
  separation of concerns.

Create dedicated form component for save note dialog:
  - Extracted note saving form logic into a separate component with improved
  tag management, form validation, and a cleaner interface for handling note
  metadata like title, category, and tags.
  - This component handles the form state and submission logic independently from
  the dialog container.
